Use the Vuex store as the primary source of reviews
Previously, the localStorage storage was the authoritative source.
Now, the Vuex store is the authoritative source, with the localStorage
store being the "backup"/persistent store.

That means we need to push the localStorage values into the app,
but reserve the `mrReviews` property for the mapped state.
We do this by renaming the initial hydrator to "rehydratedMrReviews".