Fix allow settings for maven duplicates
Prior to this change, there was an incorrect
usage of the checkbox and this caused the
test to be unable to check/uncheck properly

This change uses a button instead and the
button label to validate state. using check?
revealed a problem since it asserts for is-checked
attribute on the element being checked