Move security scan status strings to frontend
This moves the status strings from the Rails ConfigurationPresenter to
Vue components in the frontend.

This also updates the pinning tests in the previous commit. It's best to
ignore whitespace changes when viewing this commit, since there are some
indentation changes in the snapshots. For example:

    git show <this commit> --ignore-all-space

Addresses https://gitlab.com/gitlab-org/gitlab/-/issues/323375.