Fix table items, add subheader, crosslink
Several table items needed repairs to text. The information about the
two fields was buried far, far away from the parameter definition,
so I added a subheading and an internal link. Several other parameters
on this page are problematic, but I'm limiting the scope of this fix
to this one issue.