Delete notes from index that are orphaned
We discovered at
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/55878 that there
are sometimes orphaned commit notes. Since these are making it difficult
to migrate as we can't set the permissions correctly for these and also
since you can't actually navigate to a note on a deleted commit there is
no point in keeping these in the index.

This commit was extracted from a combined MR at
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/55883 so it could
be merged independently from the migration.