Handle commits without descriptions for changelogs
When processing commits to include in a changelog, we may encounter
commits without a description (= they only have a subject). This results
in Commit#description being nil, resulting in a NoMethodError when we
try to extract a revert SHA from that description.

This commit fixes this by just not matching if the description is nil.
This is fine since commits without a description are never included, due
to a lack of a trailer tag.

This fixes https://gitlab.com/gitlab-org/gitlab/-/issues/323998