Distinguish between validation errors and stale resources errors when Runner requests a new job
When job is being scheduled, GitLab tries to detect concurrent requests
that would get the same job. In that case only the first Runner that
asked for a job gets it, and all other requests are internally handled
again (until there are other jobs in the queue that matches such
specific Runner).

When such "conflict" situation was detected and no job was found in
the loop repeat, Runner receives a `409 Conflict` response, which leaves
a track about such situation. It additionally triggers a request
back-off on Runner, so in highly concurrent installations it should
generate some intervals between concurrent requests from different
Runners.

However, one of the symptoms used for finding the "conflict" situation
is the `StateMachines::InvalidTransition` exception, which may be also
thrown when some data doesn't pass validation. This is a case that
should definitely not produce a `409 Conflict` response, but instead
a `400 Bad Request` one with details about what was wrong should be
sent.

This commit adds a needed update to the `Ci::RegisterJobService`
service and the `/api/v4/jobs/request` endpoint.

More details about the problem and why we'fe decided to handle it like
that can be found at https://gitlab.com/gitlab-org/gitlab-runner/issues/4360
and especially at
https://gitlab.com/gitlab-org/gitlab-runner/-/issues/4360#note_200280465
comment.