Group#self_and_descendants -> #all_projects in GroupPackagesFinder
We have found serious performance issues when using
`#self_and_descendants` in `IN (...)` queries. We have already updated
`#all_projects` so that it does not use this anymore
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/56078 and as such
we want to change other parts of the code to just use `#all_projects`
when it makes sense to keep this performance benefit.

It also happens that this simplifies the code anyway since we actually
already wanted the list of projects in the group.