Change all Project.for_group_and_its_subgroups to group.all_projects
This is mainly motivated by performance issues we were seeing with
`WHERE ID IN(...)` queries
https://gitlab.com/gitlab-com/gl-infra/production/-/issues/3894 .

Those were addressed by updating the `#all_projects` method to use
`INNER JOIN` instead
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/56078 so it's
better if we move as much code to using `#all_projects` as possible . It
happens that `#all_projects` is actually less code to write in all these
places.

After removing all occurrences of `Project.for_group_and_its_subgroups`
it's now clear that the scope isn't even needed anymore as it was never
chained with anything else so we can remove it.