Fix ActionMailer job metric labels
When we emit metrics for Sidekiq workers, we were only adding some
labels for workers defined in our application code.
ActionMailer::MailDeliveryJob comes from Rails - we don't define that
class. But we still want to add these labels to the metrics. The urgency
label in particular is important, as it's used for calculating SLIs;
without that label, we current have no SLIs for the mailers queue.