Add Username to Email From Header in Notifications
When replying to a notification from GitLab, it can be non-obvious as
to how to ping the person I am responding to without looking for their
username in Gitlab.

By adding the username to the From header in the form of From: User
Name (@user), it is much more clear as to how to ping a person in the
response.