Remove section better documented in GitLab UI
The "Vue component interpolation" section was removed since it's not as
well written or as complete as the documentation for the GlSprintf
component in GitLab UI, and existing links to that section have been
replaced.

Also add bullet points for reasons to use GlSprintf, and a suggestion of
when sprintf is fine.