Add option for chat service label filter
This adds an option to the Slack and mattermost
chat notification services. When the option is 'true'
the issue or merge request must have applied all the
labels that are listed in labels_to_be_notified in order
to trigger a notification. If the option is 'false', a
notification will be triggered when the issue or merge request
has one of the listed labels.