Handle deprecations as a first-class concern
This makes the following changes to how we process deprecations in
our documentation:

- allow there to be common reasons
- treat replacements explicitly
- lookup the original deprecation information when processing docs

Enums are changed to use the new calling convention, but
do not record the new deprecation information yet.

Some types are changed to use the new deprecation features.

As a result of this, we need to have the full schema available
in the doc renderer, so we pass the full schema to it now, not the
graphql definition.