Include user namespace in list of fork targets
With the fork_project_form feature flag enabled, we don't pay any
attention to the `@own_namespace` instance variable, instead expecting
to get the full list of namespaces we can fork to in the JSON payload
given by Projects::ForksController#new

Adding `@own_namespace` to that json payload allows us to render it
without any special handling. It does mean that we now have a mixed
array of Namespace and Group, so a few Group-specific things need to
be supported by Namespace.