gitaly_client: Introduce new function to fetch remotes via URL
When fetching remotes, we currently first create it and then fetch the
created remote. This is an inefficient interface and awkward to use, as
we have to assert that the remote exists each time we fetch. It's also
vulnerable to races as we now depend on on-disk state of Gitaly.

Some time ago, Gitaly has introduces a new `RemoteParams` field to the
`FetchRemote()` RPC which enables callers to provide configuration of
the remote directly instead of passing the previously created remote.
As a first step towards switching to that function, this commit
introduces a new function `fetch_url()`, which is the equivalent of
`fetch_remote()` only with a URL and refmap instead of a remote name.