Add mapping for Auto DevOps testcases issues and specs
Because we're currently not linking the Auto DevOps specs to their respective testcases
issues, the issues are not tracking the pass/fail results of our test runs.  This is a boring
solution to try to get those pass/fail results into the testcases issues.  I'm not a huge
fan because I think this is hard to read, but it will hopefully at least let us keep
track of how often the tests are failing.