Create a fake mimemagic gem
Issue: gitlab.com/gitlab-org/gitlab/-/issues/325851
MR: gitlab.com/gitlab-org/gitlab/-/merge_requests/57443

There is a current issue affecting mimemagic where it was
using a magic number dataset which turned out to be under
a GPL licence, which in turn meant the maintainer of mimemagic
had to switch to the same licence. We can not use this licence.
There are still dependencies in our gemfile which require mimemagic,
and instead of removing all these, we create a fake gem that we
require.