Move after create tasks to async job
The MergeRequests::AfterCreateService was created in order to move
one problematic bit of code into an async job. It was noticed that these
items could probably be moved as well. That work was deferred to a
follow up issue to avoid slowing iteration. Since GitLab.com is
currently suffering from performance issues, this is now a priority.