Remove extend from issuable sidebar css
Makes the "+3 More" button into an actual button instead
of looking like a link. It should really look like a link
with blue text on hover, since it is more like the Expand
section buttons. Styling as a tertiary button has minimal
visible change, and small to reduce the padding changes