Automate deactivation of dormant users
Initial (and naive) implementation of automated deactivation of dormant
users with cron worker.

This new worker runs daily, and if new application setting
`deactivate_dormant_users` is enabled, iterates through all users
(using `find_each`) and deactivates those that `can_be_deactivated?`.

There is no UI for the new application setting yet.

We do not plan to ship this implementation as is but even if we do it
will work for small self-managed instances. In the next iteration we'll
improve the `User.dormant` scope to fetch actual dormant users only.

The intention is that this will never be enabled for SaaS.

See https://gitlab.com/gitlab-org/gitlab/-/issues/211754.