Remove N+1 DB queries from indexing projects in Elasticsearch
The `#as_indexed_json` method is used to convert individual documents to
the JSON that is sent to Elasticsearch during indexing. We need to
preload the data before it gets to this method. We implemented the
`.preload_indexing_data` method in
gitlab-org/gitlab!56808 so we just
need to add this to the various classes now. This MR updates it for
`Project` only.