Update fork e2e tests for new UI
This update the Fork page object to that fork tests work with the new
UI. The changes are currently behind a feature flag, however if we
updated all the tests to enable or disable the flag explicitly we'd
have to update a few tests, and then all of them would require admin
access (because we can only enable/disable feature flags via the admin
API).

Instead, this updates the page object to detect which UI elements are
shown and use the appropriate ones.

More tests will be added to explicitly test with the feature flag
enabled or disabled so that we know each state has coverage.