Split EnvironmentsFinder into two
Before this commit EnvironmentsFinder has 2 different ways of use. There
was classic execute method but there was also find method that worked
differently. This created a confusion about how to use it. After this
commit the finder split into two separate finders so we can improve or
deprecate each separately and prevent confusion.

Signed-off-by: Dmytro Zaporozhets <dzaporozhets@gitlab.com>