Set workhorse_extract_filename_base feature flag to default
Before this feature flag was enabled, Workhorse would reject any
multipart filename that contained a slash (e.g. `dist/release.tar.gz`),
even though GitLab never used the path in determining whether to store
the file. With this feature flag enabled, Workhorse extracts the base
from this filename so this path is effectively ignored. This allows
uploads via the API to work as before the endpoint was accelerated by
Workhorse.

Relates to:

1. https://gitlab.com/gitlab-org/gitlab/-/issues/326379
2. https://gitlab.com/gitlab-org/gitlab/-/issues/326350