Fix N+1 queries to find or initialize services
Find or initialize the services has the following precedence:
- Find records.
- Build services them from instance-level record or service template
  record.
- Build services from scratch.

The problem this fixes when building the services from scratch, Rails
method build_(association) performs a query for each service that we
have to initialize.

So for each service that we build we were doing one query. Now using
`NameService.new` we reduce the query count from 38 to 4.