Limit number of GraphQL requests tracked in performance bar to 10
Previously, we only tracked up to 2 requests to the same URL in the
performance bar. For GraphQL, we removed that limit as the URL is the
same for all GraphQL requests, and we didn't want to miss anything that
was genuinely new.

However, with no limit, if we poll using GraphQL then we'll end up with
this store increasing forever. So we set a higher - but still reasonable
- limit of 10 GraphQL requests to be stored in the performance bar.
Other requests can still be added manually by request ID as normal.