Reset state correctly for tests failed in Rails 6.1
Objects created via let_it_be are shared between tests.
Sometimes we reload them in order to reset the state.
Sometimes it's not enough.
For example, some tests fail because @_destroyed instance
variable preserved. So the objects removed in a previous
test, marked as removed in the subsequent ones.

Let's use refind to fix those cases