Change modal error message for restricted emails
Displays a different error message for  the API error message returned
when an email address is invited that is not on the restricted domain list, when this EE only
feature is enabled. The message is different depending on whether it is an invite to a group
or to a project, and whether the page viewer is able to administer members or not.