Extract commit count logic to use elsewhere
Extract commit count logic used in `new_project_readme` experiment for
use in `empty_repo_upload` experiment.

It came to our attention via the Data Analysis team that the
`empty_repo_upload` tracking for `initial_commit` was too naive, and
wasn't working as intended, but a similar event for the
`new_project_readme` experiment was working as expected.

I decided to extract the method of retrieving commit counts into a
simple shared interface.

* adds a EmptyRepoUploadExperiment class which doesn't introduce the
  experiment, but more so the experiment now has it's own class and is
  no longer a base class experiment

part of:
https://gitlab.com/gitlab-org/gitlab/-/issues/325974